Create the Group chat feature without modify
Now, continue on the group chat feature, first let the admin be able to create a group chat. When you click on the 'group_add' Icon, it will navigate to a page to fill in the group information, which includes the group name, group image, and allows you to key in the user ID to add to the group. It will check whether the user_id exists or not.

Can have a group chat ID to let the user key in to join the group.

For the chat list page of the user, add one more feature on top of the app bar, which is join group, which allows the user to join a group when the user key in the group chat ID. After key in the group chat ID, it will send the request to the admin to request to join the group.

TODO:
The admin is able to manage the request at the group chat page. At the top left of the group chat page, you will have the settings of the group chat, which allow the admin to manage the group information. When you click on the settings icon, it will navigate to the page that manages the information of the group chat. The admin is able to change the group name, group image, add or delete users, and also manage the requests, which can be rejected or accepted the request.