continue on doing the group chat feature and also broadcast message
Now, i want to implement the admin management for the group chat. It will display a manage group page to let the admin modify the group information, which include change the group name and group image. It also allow the admin to add or remove the group member. When there are users request to join the group chat, the request will also be display at the manage group page.

Group Chat Page: Admin Management
Settings Icon:
Top right of the group chat page, visible only to admin.
Navigates to a "Manage Group" page.

Manage Group Page:
Editable group name and image
List of members (with remove option)
Add icon with word 'add mamber' (use to add member)
List of join requests (with accept/reject buttons)

Admin can:
Change group name and image.
Add or remove users.
See join requests, and accept or reject them (move from requests to members or remove from requests).

Then I implement the Broadcast Message feature. It’s one-way communication from the admin to all users, which is used to quickly inform all users of critical updates. The message history will be automatically pinned at the top of the chat list to let users quickly see the message.

Admin writes a broadcast message and can optionally attach an image and location. Displays a message in a read-only broadcast section.

The appbar of the chat page for the broadcast message can be 'Announcements'. The message can be listed with an Urgency label (Critical / Warning / Update), Timestamp and the Message text. It may also be attached to an image and a location. There can be different colours to represent the Urgency label, which are red for Critical, orange for Warning and blue for normal information.

TODO: Test the feature of group chat and broadcast